// Client/SpiskerApp/ClientApp/app/Social/Model/UserProfileModel.ts
export class UserProfileModel {
    userName: string;

    email: string;

    firstName: string;

    lastName: string;

    middleName: string;

    statusText: string;

    avatarUrl: string;

    listCount: number;

    followerCount: number;

    followingCount: number;

    isFollowing:boolean;
}